Make payments listen on additional http port
And remove ApplePay specific customizations. This way payments is
always listening on an https port (9001 by default as before) and
an http port (9009 by default). We set $wgServer in LocalSettings
by calling a function that examines the $_SERVER request variables
so it's valid for either, and even for requests coming in on a
totally different hostname.

Depends on an update to the payments image:
https://gitlab.wikimedia.org/releng/dev-images/-/merge_requests/8

With that image update and this patch, you should be able to
forward one of the 8001-8006 ports from payments.fr-tech-dev to
your port 9009 with this command:

ssh -R 8001:localhost:9009 payments.fr-tech-dev

Then docker payments should respond to requests to
https://paymentstest1.wmcloud.org

Bug: T291577
Change-Id: I1875988f74e39223ddd6251dbf185f231707e737