Add service for donatewiki and centralnotice
Helpfully named 'donut'

Uses a new version of the fundraising-mediawiki-apache2 image just
published by releng (no need to build your own).

That image uses a site config file from the host config directory which
is included in this patch. Note that it differs from the site config
previously included in the image, adding a RewriteRule for /w/ URLs
since DonationForm.js is loaded that way. It also deletes some rules
that had no effect because they were not under a directory or vhost tag.

A dump of Donatewiki content as of 2022-12-07 is included gzipped.
It is lightly edited on import to replace links to production
payments with links to our local payments container.

Note I had to add -O to the git hook scp command as suggested here:
  https://github.com/PowerShell/Win32-OpenSSH/issues/1945
to avoid the 'subsystem request failed on channel 0' error. I'll
probably need to add it to the rest of the similar commands if I
ever want to reclone.

CentralNotice is included along with Translate and
FundraisingTranslateWorkflow. UniversalLanguageSelector is also
included as a dependency of Translate. There is some minimal
configuration for CentralNotice and the fundraising workflow.

Maybe TODO: import some of the banner JS and CSS from mediawiki?

Bug: T317498
Change-Id: I017427116382e60508e6599dd5aeb4fc808cdcb5