Donut: Enable JS tests and remove setting
In addition to enabling local JS QUnit tests by setting
$wgEnableJavaScriptTest, we remove
$wgCentralNoticeImpressionEventSampleRate, which was interfering
with one test.

(We should also improve that test, but also it seems
preferable to have this setting locally set to the same value
as production, which is what we get with this patch.)

To test, go to Special:JavaScriptTest.

Documentation: https://www.mediawiki.org/wiki/Manual:JavaScript_unit_testing

Bug: T326974
Change-Id: Ia69dd737d481b1dada53cb068fe0f818653f1294