Donut: Use /w for $wgScriptPath
This also moves the base install directory to /w under
the Apache DocumentRoot inside the container, and sets
config to allow the /wiki path for simple page names.

This follows production configuration for both Donate
Wiki and other production wikis that run CentralNotice.

With this patch, you should be able to go directly to a
page with the "wiki" short URL, for example:
https://localhost:9010/wiki/Special:FundraiserRedirector.

Links in content imported directly from Donate Wiki should
also now work, including the special "Tools" sidebar.

Bug: T326974
Change-Id: Ia28d4cc77faf3be660d314b6f9fad4c0ed7a4f3a