Add more queue consumers
And settings for more queues.

Both settings and consumer for upi-donations (see Ibc09240bffdbd95a12
and I92e33a41aa624674c01522) and jobs-dlocal (see Ida30895a88934a3df1),
and consumer for email-preferences.

Also ensure we are running like we do in production by adding @wmff
and increasing the vebosity to -vv. Note that with -vv we get a big
annoying 'Box Requirements Checker' block at the start of each drush
job, but we need that extra verbosity for the \Civi::log() statements
eillen is replacing watchdog with to show up.

Change-Id: I315c79527d4ac1a05437e84dcc07621c73ac9a03