Pass through site_key in config copy of civiproxy
We have to make this change twice, once in the 'real' copy of proxy.php
and once in this copy which has the strict cert checking stripped out
to allow us to run civiproxy on our dev instances. See the Depends-On
patch for the related changes to the rest of civiproxy.

The code in civicrm core that examines these params is in the authx
extension which now ships with core:
https://phabricator.wikimedia.org/diffusion/WFCG/browse/master/drupal/sites/all/modules/civicrm/ext/authx/Civi/Authx/Authenticator.php$116

Note that you will have to make sure your admin@example.com civi user
has a row in civicrm_uf_match pointing to the corresponding drupal
user. Mine was missing for some reason but after a rebuild it was back.

Bug: T349946
Change-Id: I124ccfb98aba4c279668a2813251d82c29efb70b
Depends-On: I96a85540c1af669d1b7c3c9fbfa7760488dc73ca