Rename queue helpers to all use a queues-function convention.
We don't need the redis bit here as there's no other
queue backends so it's assumed to always be redis.

Catalyst: I was trying to monitor the queues and forgot
the redis bit and then realised we probably don't need it.
We also don't use it for the get & set ct-id scripts.

Change-Id: I6473573b444b811ef18a9d4dd1c79fc113e41f3f